+++ test/unit/index.js
@@ -150,51 +150,34 @@ describe("index: options", function () {
     
   });
 
-  describe(".jsCompilers", function () {
+  describe(".compilers", function () {
 
     it("supports javascript by default", function () {
       var options = connectAssets.parseOptions({});
-      expect(options.compilers.js.js).to.be.ok();
+      expect(options.compilers.js).to.be.ok();
     });
 
     it("supports coffeescript by default");
-
-    it("allows adding a new compiler without replacing defaults", function () {
-      var options = connectAssets.parseOptions({ jsCompilers: { ts: {} } });
-
-      expect(options.compilers.js.js).to.be.ok();
-      expect(options.compilers.js.ts).to.be.ok();
-    });
-
-    it("allows overriding default compilers", function () {
-      var options = connectAssets.parseOptions({ jsCompilers: { js: "e" } });
-
-      expect(options.compilers.js.js).to.be("e");
-    });
-
-  });
-
-  describe(".cssCompilers", function () {
 
     it("supports stylus by default");
 
     it("supports less by default", function () {
       var options = connectAssets.parseOptions({});
 
-      expect(options.compilers.css.less).to.be.ok();
+      expect(options.compilers.less).to.be.ok();
     });
 
     it("allows adding a new compiler without replacing defaults", function () {
-      var options = connectAssets.parseOptions({ cssCompilers: { sass: {} } });
+      var options = connectAssets.parseOptions({ compilers: { ts: {} } });
 
-      expect(options.compilers.css.css).to.be.ok();
-      expect(options.compilers.css.sass).to.be.ok();
+      expect(options.compilers.js).to.be.ok();
+      expect(options.compilers.ts).to.be.ok();
     });
 
     it("allows overriding default compilers", function () {
-      var options = connectAssets.parseOptions({ cssCompilers: { css: "e" } });
+      var options = connectAssets.parseOptions({ compilers: { js: "e" } });
 
-      expect(options.compilers.css.css).to.be("e");
+      expect(options.compilers.js).to.be("e");
     });
 
   });